python_watchlist_gui/requirements.txt
websocket-client==1.6.2
requests>=2.31.0
# tkinter is included with most Python installs; no pip package required
pytest>=7.0
